Create HighlightDeprecatedHTTPMethods.bambda
This code targets HTTP requests that use deprecated or less common methods, such as TRACE and CONNECT. These methods are often overlooked but can be exploited in certain types of network attacks. When such a request is detected, the script highlights it in red within the Burp Suite interface, making it easy for security analysts to spot and investigate these potentially risky requests.